backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiohttp==3.9.1
pydantic==2.5.0
reportlab==4.0.7
pyzbar==0.1.9
pillow==10.1.0
